Backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==65.5.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docutils==0.21.2
drf-yasg==1.21.7
id==1.5.0
idna==3.10
inflection==0.5.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.6.0
nh3==0.2.21
packaging==24.1
phonenumbers==8.13.45
pillow==10.4.0
Pygments==2.19.1
PyJWT==2.9.0
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
setups==0.0.2
setuptools==78.1.0
sqlparse==0.5.1
stripe==12.0.0
twine==6.1.0
typing_extensions==4.13.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.45.1